+++ src/main/java/des/SBox.java
@@ -16,7 +16,7 @@ public class SBox extends BitBlock {
     }
 
     public static byte getBoundsOfSix(byte sixBitInput) {
-        byte rightBound = (maskBits(sixBitInput, BitMask.LASTBIT));
+        byte rightBound = (byte)(maskBits(sixBitInput, BitMask.LASTBIT));
         byte leftBoundShifted = (byte) (maskBits(sixBitInput, BitMask.THIRDBIT) >> 4);
         return (byte) (rightBound | leftBoundShifted);
     }
